+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.0.1] 21. January 2019
 
-
 ### Added
 - Adapting semating versioning
 - Spool: added log control
@@ -14,3 +13,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Log control
 - Documentation fixes and internal references.
+- Fixed Gamma import from Lat/Lon